Module Jsonxt.Extended_monad

Extended_monad supports parsing and writing JSON data that conforms to the Json.Extended.json json type using reader and writer funtions that implement the IO monad. This supports non-standard JSON types including integer as well as tuples and variants introduced by Yojson

module type IO = Jsonxt__.Io.IO

module Make : functor (IO : IO) -> sig ... end

The Make functor is used to create a module with reader and writer functions using the IO monad.