Module Jsonxt__.Basic_monad

module type IO = Jsonxt__.Io.IO

module Make : functor (IO : IO) -> sig ... end

The Make functor is used to create a module with reader and writer functions using the IO monad.