Module type Process_intf.Basic

type json

val to_number : [> `Int of int | `Float of float ] -> float

to_number json converts `Float f to f and Int i to float i

val to_number_option : [> `Int of int | `Float of float | `Null ] -> float option

to_number_option json converts `Float f to Some f, `Int i to Some (float i) and `Null to None

val to_int : [> `Int of int ] -> int

to_int json converts `Int i to i

val to_int_option : [> `Int of int | `Null ] -> int option

to_int_option json converts `Int i to Some i and `Null to None

val filter_int : [> `Int of int ] list -> int list

filter_int l returns a list of all the values of `Int value elements in l

val filter_number : [> `Int of int | `Float of float ] list -> float list