Module type Jsonxt__Process_intf.Extended

type json

val sort : [> `Assoc of (string * 'a) list | `List of 'a list | `Tuple of 'a list | `Variant of 'b * 'a option ] as 'a -> 'a

sort json sorts the json tree based on field names. Objects and lists are sorted recursively. Note that the function only sorts field names and not the values. The sort is stable