Module Jsonxt__Strict_monad

Strict_monad supports parsing and writing JSON data that conforms to the Json.Strict.json json type using reader and writer funtions that implement the IO monad. This only supports types supported by the JSON standard and explicity excludes integers

module type IO = Jsonxt__.Io.IO

module Make : functor (IO : IO) -> sig ... end

The Make functor is used to create a module with reader and writer functions using the IO monad.